Fix bug where auto-login could happen but still warn credentials were expired
Specifically, this issue arose when `_silent_login` determined that
you were using an outdated/deprecated .onecodex file and warned you,
even though the Api() instantiation could parse it. Fix is to explicitly
pass '' as the `api_key` rather than `None`. We'll likely need to do the
same when we switch to default JWT tokens for the CLI.