Simplify upload handling (#246)
* Simplify upload code -- temporarily reverting to single threaded model

* Support multi-threaded uploads via `concurrent.futures`

This commit also reworks the passing of the click.progressbar
to the uploader and requires it's passed as an object vs. boolean
from the top invoking call. This better supports both library and CLI
usage.

* Implement custom progress bar, cleanup utils.py

Specifically we add support for "canceling" and "finalizign" statuses
in the click.progressbar element. Note that we could generalize this
further by subclassing the click internal class, but it's already a
sufficiently gross hack that we're probably better off *not* further
extending the click internals.

* Implement robust upload cancelation for SIGINT and SIGTERM incl across threads

We do this via the use of the `atexit` module, which will handle any
program exits. Note we have to use our own `atexit_unregister` function
because `atexit.unregister` is unavailable in Python 2.7.

* Simplify exception handling logic in lib/upload.py

Notably, remove the use of boolean return values and rely exclusively
on exception hierarchy and retry-able vs. non-retryable exceptions.
This includes minor flow control logic changes (exceptions vs. return
values), but no major changes to the upload code paths.

* Overhaul and simplify proxy upload routines, handling ConnectionErrors

Effectively, we do the following here:
1. Always check for /status if available when uploading via a proxy
2. If /status has a `complete: false` field, poll up to 15 minutes
3. Raise a RetryableUploadException or fail based on /status code

(2) specifically fixes bugs we were encountering wherein the Python
client could send all of the data and either:
A. Send a file with an error but not be listening for the reply,
   experience a ConnectionError, and then retry automatically
   even if the file was invalid and the proxy returned a 400
B. Send data too fast for the proxy to compress (with data getting
   buffered in the ALB/nginx) and then abort *despite the proxy
   ultimately completing successfully*

(A) led to bad files not being caught by the proxy. (B) could lead to
duplicate uploads to to erroneous retry requests. Both of these fixes
bring our implementation into line with the frontend JS code on
app.onecodex.com.

* Improve formatter for error vs. info logging messages

No longer display "INFO" for logging.info messages when invoked via the
CLI. We can now further configure these for both Python 2 and Python 3.