Faster imports for `viz` and `analyses` subpackages
Sped up import times for `onecodex.viz` and `onecodex.analyses` subpackages. Testing locally on a macOS laptop, the previous import times for both packages were ~670ms, and the new import times are ~170ms.

Faster imports are achieved by defining an import hook to configure Altair's theme and renderer the first time it is imported. Directly importing and configuring Altair in the `onecodex.viz` subpackage was the primary cause of the slow imports. An import hook avoids this performance hit by configuring Altair during deferred import in visualization code.

I also deferred a few more imports in the `viz` subpackage, and added a couple of unit tests to ensure Altair is being configured as expected.

Addresses part of #248.