Replacing chao1 with observed_taxa (#360)
* Replacing chao1 with observed_taxa

Chao1 is not an ideal metric for whole metagenome shotgun data. Observed taxa is a more sensible metric.

* allowing chao1 while warning of deprecation

We will still support Chao1 in this version. This commit adds Chao1 back after I removed it in another commit. However if a user calls on Chao1, we will now warn that it is deprecated and will be removed.

* removing superfluous parentheses and including warnings

I had errors in my previous commit, which included an extra ) and missing "import warnings".
I've also run the pre-commit hooks.

* Implementing review changes

@jairideout and @clausmith proposed changes during review. These are now implemented.