Add download retries to ResourceDownloadMixin (#361)
* Add download retries to ResourceDownloadMixin

Added retries to `ResourceDownloadMixin` to support downloading files that may not be immediately available (e.g. downloading an `Assembly` FASTA file that is exported and uploaded to S3 asynchronously).

Also added support for using the server's suggested "save as" filename if the user doesn't provide a path or file object. Currently, only the `Assemblies` resource uses this new feature.

This change is necessary to support https://github.com/onecodex/mainline/pull/4057

* Add a few coverage tests for ResourceDownloadMixin

* Update code to match latest `mainline` API

* Download progress bar: use download path's filename if available, otherwise use resource's filename