Improve missing value coercion in SampleCollection._collate_results (#410)
Falsey values are now explicitly set to zero in `SampleCollection._collate_results()` instead of relying on numpy + pandas to handle it. This wasn't actually an issue because `None` converts to `np.nan` in a float dtype array, which was then converted to zero. An error would have been raised if there were missing values for an int dtype metric though. The new code handles both cases.