Support missing fields in AnalysisMixin._metadata_fetch()
Updated `AnalysisMixin._metadata_fetch()` to support fields that match neither metadata nor taxa. Instead of raising an error, the field is filled with `None` to represent an empty column of data.

This can be useful, for example, when working with metadata on a `SampleCollection` that has been filtered and the field no longer exists on the filtered collection.

Closes DEV-6191
Closes DEV-6700