Add support for linking to NCBI taxonomy browser from plots
`plot_bargraph()` and `plot_heatmap()` have a new `link` parameter for controlling whether users are linked to the sample's classification results in the One Codex app (default), or to the NCBI taxonomy browser for the taxon that was clicked on.

Also, links are now opened in a new tab/window for all 6 plotting methods.

Closes DEV-7213