Add secondary haxis grouping to `plot_metadata()`
Added optional `secondary_haxis` parameter to `plot_metadata()`, which can be a field name or tuple of field names.

If provided, a grouped scatterplot or boxplot is generated, where `haxis` is the primary grouping and `secondary_haxis` is the secondary grouping. The secondary grouping is colored and includes a legend.

Also DRYed up unit tests relying on a specific `SampleCollection`.

Closes DEV-9216