Upgrade to Altair 5 and replace altair_saver with vl-convert (#507)
Upgraded to Altair 5.4.1 and replaced altair_saver with vl-convert.

vl-convert is a Rust package with Python bindings and no external dependencies. altair_saver is no longer maintained and is only supported in Altair 4 and earlier. node and vega npm packages are no longer required to render PNG, SVG, and PDF formats.

Also:

- Fixed Jupyter Notebook install by upgrading to `notebook` 6.5.7
- Removed jupyter-client dependency
- Removed selenium dependency