Filter and warn if samples are missing abundances in `plot_metadata()`
Samples with missing abundances are filtered out in `plot_metadata()` when plotting alpha diversity or taxon abundances. A warning is also raised.

If numeric metadata is plotted on the vertical axis, this filtering is not applied because abundances are not needed to create the plot.

An error is raised if all samples are filtered out due to missing abundances.

Closes DEV-10065